requirements.txt
greenlet==2.0.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
